requirements.txt
flask~=3.0.3
gunicorn
google-generativeai
python-dotenv
protobuf~=4.25.3